Breakout logic in start to individual files per component
